correct errors in units of hbar/m

+++ src/enzo/Grid_ZeusSolver.C
@@ -231,7 +231,7 @@ int grid::ZeusSolver(float *gamma, int igamfield, int nhy,
 	ENZO_FAIL("Error in CosmologyComputeExpansionFactors.");
       }
     // calculate hbar/m
-    float hmcoef = 5.9157166856e27*TimeUnits/pow(LengthUnits/a,2)/FDMMass;
+    double hmcoef = 5.9157166856e27*TimeUnits/pow(LengthUnits/a,2)/FDMMass;
     //(hbar/m)^2/2
     double lapcoef = POW(hmcoef,2)/2.;
